shared/km: kernel mode debug logging
This commit adds kernel mode debug logging for FUSE requests/responses.

In addition this commit fixes a bug in FuseInstanceTransact that was
discovered by this new logging. FuseInstanceTransact would zero out only
the header of a FuseRequest, which could leave stale data in the remainder
of the FuseRequest, which could then confuse libfuse. The fix was to ensure
that the whole FuseRequest is zeroed out.